requirements.txt
altair
pandas
streamlit
plotly-express
openpyxl
psycopg2-binary